# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Salli API",
    "description": "Privacy-first personal finance, tax and financial independence",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/subscriptions/reports": {
      "get": {
        "tags": [
          "subscriptions"
        ],
        "summary": "Get All Reports",
        "description": "Missed-charge/price-change reports for every active subscription.",
        "operationId": "subscriptions.reports",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/mcp/oauth/consent": {
      "post": {
        "tags": [
          "mcp-oauth"
        ],
        "summary": "Consent",
        "operationId": "oauth.consent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConsentDecisionRequest": {
        "properties": {
          "rt": {
            "type": "string",
            "title": "Rt"
          },
          "approve": {
            "type": "boolean",
            "title": "Approve"
          }
        },
        "type": "object",
        "required": [
          "rt",
          "approve"
        ],
        "title": "ConsentDecisionRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}